+++ city_metrix/layers/fab_dem.py
@@ -7,12 +7,12 @@ from ..constants import GTIFF_FILE_EXTENSION
 DEFAULT_SPATIAL_RESOLUTION = 30
 DEFAULT_RESAMPLING_METHOD = 'bilinear'
 
-class NasaDEM(Layer):
+class FabDEM(Layer):
     OUTPUT_FILE_FORMAT = GTIFF_FILE_EXTENSION
     MAJOR_NAMING_ATTS = None
     MINOR_NAMING_ATTS = None
 
-    def __init__(self,  **kwargs):
+    def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
     """
@@ -22,12 +22,12 @@ class NasaDEM(Layer):
     """
     def get_data(self, bbox: GeoExtent, spatial_resolution:int=DEFAULT_SPATIAL_RESOLUTION,
                  resampling_method:str=DEFAULT_RESAMPLING_METHOD):
-
+        
         spatial_resolution = DEFAULT_SPATIAL_RESOLUTION if spatial_resolution is None else spatial_resolution
         resampling_method = DEFAULT_RESAMPLING_METHOD if resampling_method is None else resampling_method
         validate_raster_resampling_method(resampling_method)
 
-        nasa_dem = ee.Image("NASA/NASADEM_HGT/001")
+        fab_dem = ee.ImageCollection("projects/sat-io/open-datasets/FABDEM")
 
         ee_rectangle  = bbox.to_ee_rectangle()
 
@@ -35,9 +35,9 @@ class NasaDEM(Layer):
         kernel = ee.Kernel.gaussian(
             radius=3, sigma=1, units='pixels', normalize=True
         )
-        nasa_dem_elev = (ee.ImageCollection(nasa_dem)
+        fab_dem_elev = (ee.ImageCollection(fab_dem)
                          .filterBounds(ee_rectangle['ee_geometry'])
-                         .select('elevation')
+                         .select('b1')
                          .map(lambda x:
                               set_resampling_for_continuous_raster(x,
                                                                    resampling_method,
@@ -50,13 +50,13 @@ class NasaDEM(Layer):
                          .mean()
                          )
 
-        nasa_dem_elev_ic = ee.ImageCollection(nasa_dem_elev)
+        fab_dem_elev_ic = ee.ImageCollection(fab_dem_elev)
         data = get_image_collection(
-            nasa_dem_elev_ic,
+            fab_dem_elev_ic,
             ee_rectangle,
             spatial_resolution,
-            "NASA DEM"
-        ).elevation
+            "FAB DEM"
+        ).b1
 
         # Round value to reduce variability
         rounded_data = data.round(2)