+++ city_metrix/layers/natural_areas.py
@@ -2,12 +2,12 @@ from xrspatial.classify import reclassify
 
 from city_metrix.metrix_model import Layer, GeoExtent
 from .esa_world_cover import EsaWorldCover, EsaWorldCoverClass
-from ..constants import GTIFF_FILE_EXTENSION
+from ..constants import GTIFF_FILE_EXTENSION, DEFAULT_PRODUCTION_ENV
 
 DEFAULT_SPATIAL_RESOLUTION = 10
 
 class NaturalAreas(Layer):
-    GEOSPATIAL_FILE_FORMAT = GTIFF_FILE_EXTENSION
+    OUTPUT_FILE_FORMAT = GTIFF_FILE_EXTENSION
     MAJOR_NAMING_ATTS = None
     MINOR_NAMING_ATTS = None
 
@@ -20,7 +20,7 @@ class NaturalAreas(Layer):
             raise Exception('resampling_method can not be specified.')
         spatial_resolution = DEFAULT_SPATIAL_RESOLUTION if spatial_resolution is None else spatial_resolution
 
-        esa_world_cover = EsaWorldCover().get_data_with_caching(bbox, spatial_resolution=spatial_resolution)
+        esa_world_cover = EsaWorldCover().get_data_with_caching(bbox=bbox, s3_env=DEFAULT_PRODUCTION_ENV, spatial_resolution=spatial_resolution)
         reclass_map = {
             EsaWorldCoverClass.TREE_COVER.value: 1,
             EsaWorldCoverClass.SHRUBLAND.value: 1,